fix event code

+++ main-service/main-dto/src/main/java/ru/practicum/dto/event/EventFullDto.java
@@ -1,11 +1,11 @@
 package ru.practicum.dto.event;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import ru.practicum.dto.category.CategoryDto;
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
+import ru.practicum.dto.category.CategoryDto;
 import ru.practicum.dto.event.enums.State;
 import ru.practicum.dto.user.UserShortDto;
 
@@ -17,7 +17,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class EventFullDto {
-    int id;
+    Long id;
     @NotBlank
     String annotation;
     CategoryDto category;
