+++ main-service/main-dto/src/main/java/ru/practicum/dto/event/UpdatedEventDto.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class UpdatedEventDto {
     String annotation;
-    int categoryId;
+    long categoryId;
     String description;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     LocalDateTime eventDate;
